fix spark test
